Fixed Sprite width and height uniform reading in sprite.frag

+++ src/main/scala/engine/render/shaders/fragment/sprite.frag
@@ -4,14 +4,14 @@ in vec2 vPos;
 
 uniform vec4 uColor;
 uniform sampler2D uTexture;
-uniform int uWidth;
-uniform int uHeight;
+uniform float uWidth;
+uniform float uHeight;
 
 out vec4 FragColor;
 
 void main()
 {
-	// FragColor=vec4(pixelIndex,0.,1.);
-	// FragColor=vec4(pixelIndex/vec2(uWidth,uHeight),0.,1.);
+	// FragColor=vec4(vPos,0.,1.);
+	// FragColor=vec4(vPos/vec2(uWidth,-uHeight),0.,1.);
 	FragColor=texture(uTexture,vPos/vec2(uWidth,-uHeight));
 }
